backend/requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.18

# Firebase
firebase-admin==6.5.0

# Data Validation
pydantic==2.10.5
email-validator==2.2.0

# CORS and Middleware
starlette==0.42.0

# Date and Time
python-dateutil==2.9.0

# Security
cryptography==44.0.0
pyjwt==2.10.1
passlib[bcrypt]==1.7.4

# Development Tools
pytest==8.3.4
black==24.10.0
isort==5.13.2
flake8==7.1.1

# Optional utilities
requests==2.32.3
typer==0.15.1